pyproject.toml
[tool.poetry]
name = "song-vibe"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
sentence-transformers = "^2.5.1"
streamlit = "^1.31.1"
numpy = "^1.26.4"
faiss-cpu = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
